Handle output and move types
