+++ cmd/goreorder/commands.go
@@ -15,30 +15,20 @@ import (
 	"github.com/spf13/viper"
 )
 
-const (
-	usage = `%[1]s reorders the types, methods... in a Go
-source file. By default, it will print the result to stdout. To allow %[1]s
-to write to the file, use the -write flag.`
+var (
+	version                     = "master"  // changed at compilation time
+	defaultOutpout    io.Writer = os.Stdout // default output is stdout
+	defaultErrOutpout io.Writer = os.Stderr // default error output is stderr
 )
 
-var (
-	version  = "master" // changed at compilation time
-	examples = []string{
-		"$ %[1]s reorder --write --reorder-types --format gofmt file.go",
-		"$ %[1]s reorder --diff ./mypackage",
-		"$ cat file.go | %[1]s reorder",
-	}
-	completionExamples = []string{
+func buildCompletionCommand() *cobra.Command {
+	var completionExamples = []string{
 		"$ %[1]s completion bash",
 		"$ %[1]s completion bash -no-documentation",
 		"$ %[1]s completion zsh",
 		"$ %[1]s completion fish",
 		"$ %[1]s completion powershell",
 	}
-	defaultOutpout io.Writer = os.Stdout
-)
-
-func buildCompletionCommand() *cobra.Command {
 	noDocumentation := false
 	bashv1Completion := false
 	completionCmd := &cobra.Command{
@@ -80,6 +70,14 @@ func buildCompletionCommand() *cobra.Command {
 }
 
 func buildMainCommand() *cobra.Command {
+	const usage = `%[1]s reorders the types, methods... in a Go
+source file. By default, it will print the result to stdout. To allow %[1]s
+to write to the file, use the -write flag.`
+	var examples = []string{
+		"$ %[1]s reorder --write --reorder-types --format gofmt file.go",
+		"$ %[1]s reorder --diff ./mypackage",
+		"$ cat file.go | %[1]s reorder",
+	}
 
 	cmd := cobra.Command{
 		Use:     "goreorder [flags] [file.go|directory|stdin]",
@@ -94,6 +92,10 @@ func buildMainCommand() *cobra.Command {
 			return fmt.Errorf("You need to specify a command or an option")
 		},
 	}
+
+	// my god this is so cool...
+	cmd.SetOut(defaultOutpout)
+	cmd.SetErr(defaultErrOutpout)
 
 	config := &ReorderConfig{
 		FormatToolName: "gofmt",
